Heliotat-Segmentation-Project/requirements.txt
flask
streamlit
requests
pandas
openpyxl
torch
torchvision
pillow
ultralytics
